리뷰별점구현

+++ t09ether/src/main/java/com/t09ether/home/dao/OnlineReviewDAO.java
@@ -7,7 +7,7 @@ import com.t09ether.home.dto.OnlineReviewDTO;
 public interface OnlineReviewDAO {
 	
 	public List<OnlineReviewDTO> selectReview(int pro_code);
-	public int rateAvg(int pro_code);
+	public double rateAvg(int pro_code);
 	public List<String> userList(int pro_code);
 	public int reviewInsert(OnlineReviewDTO dto);
 	public int reviewDel(int on_r_no);
